# sdks/sdkwork-agent-internal-sdk/openapi/sdkwork-agent-internal-api.openapi.yaml
openapi: 3.1.2
info:
  title: SDKWork Agent Internal API
  version: 0.1.0
  description: Application-ingress internal runtime API for SDKWork agent runtime hosts and trusted in-app automation.
...
servers:
  - url: '{applicationPublicIngressOrigin}'
    description: Runtime origin from topology env `SDKWORK_KERNEL_APPLICATION_PUBLIC_HTTP_URL`.
    variables:
      applicationPublicIngressOrigin:
        default: http://127.0.0.1:18280
        description: Local default from `standalone.development`; production profiles override via topology env files.
...
paths:
...
  /internal/v3/api/intelligence/runtime/snapshot:
    get:
      x-sdkwork-owner: sdkwork-kernel
      x-sdkwork-api-authority: sdkwork-agent-internal-api
      tags: [intelligence]
      summary: Load runtime aggregate snapshot
      operationId: runtime.snapshot.load
      security:
        - ApiKey: []
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: internal-api
      x-sdkwork-auth-mode: ingress-token
      x-sdkwork-domain: intelligence
      x-sdkwork-resource: runtime
      x-sdkwork-tenant-scope: application
      parameters:
        - $ref: '#/components/parameters/SdkworkTenantIdHeader'
        - $ref: '#/components/parameters/SdkworkUserIdHeader'
        - $ref: '#/components/parameters/SdkworkIdentityMacHeader'
      responses:
        '200':
          description: Runtime aggregate snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSnapshotResponse'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Application ingress token. Runtime also accepts Authorization Bearer and x-sdkwork-access-token.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT ingress when SDKWORK_KERNEL_INGRESS_AUTH_MODE=jwt (HS256, RS256, or JWKS-backed keys).
  parameters:
...
    SdkworkTenantIdHeader:
      name: x-sdkwork-tenant-id
      in: header
      required: false
      schema:
        type: string
      description: Caller tenant identity for session ownership in token ingress mode.
    SdkworkUserIdHeader:
      name: x-sdkwork-user-id
      in: header
      required: false
      schema:
        type: string
      description: Caller user identity for session ownership in token ingress mode.
    SdkworkIdentityMacHeader:
      name: x-sdkwork-identity-mac
      in: header
      required: false
      schema:
        type: string
      description: HMAC-SHA256 hex proof over "{tenantId}\n{userId}" keyed by the ingress token; required for non-loopback signed identity mode.
...
  responses:
    Problem:
      description: RFC 9457 problem detail response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    ProblemDetail:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          description: Request endpoint occurrence in the form `{METHOD} {routeTemplate}`.
        operationId:
          type: string
        code:
          type: integer
          format: int32
          description: Numeric error result code. MUST be non-zero. See API_SPEC.md §15.3.
          minimum: 40001
          maximum: 79999
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
        errorCategory:
          type: string
        retryable:
          type: boolean
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required: [code, data, traceId]
      properties:
        code:
          type: integer
          format: int32
          description: Numeric success result code. MUST be 0 on HTTP 2xx JSON bodies.
          enum: [0]
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
...
    SdkWorkResourceData:
      type: object
      additionalProperties: false
      required: [item]
      properties:
        item:
          description: Typed domain resource for the operation.
...
    Capability:
      type: object
      required: [capabilityId, providerId, status, required]
      properties:
        capabilityId:
          type: string
        providerId:
          type: string
        status:
          type: string
        required:
          type: boolean
...
    RuntimeSnapshot:
      type: object
      required: [runtime, events, permissions, workspace, patches, verificationReports, terminalCommands, terminalOutput, reviewFindings]
      properties:
        runtime:
          $ref: '#/components/schemas/KernelRuntime'
        events:
          type: array
          items:
            $ref: '#/components/schemas/KernelEvent'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionRequest'
        workspace:
          $ref: '#/components/schemas/Workspace'
        patches:
          type: array
          items:
            $ref: '#/components/schemas/PatchSet'
        verificationReports:
          type: array
          items:
            $ref: '#/components/schemas/VerificationReport'
        terminalCommands:
          type: array
          items:
            $ref: '#/components/schemas/TerminalCommand'
        terminalOutput:
          type: array
          items:
            $ref: '#/components/schemas/TerminalOutputChunk'
        reviewFindings:
          type: array
          items:
            $ref: '#/components/schemas/ReviewFinding'
    KernelRuntime:
      type: object
      required: [runtimeId, agentId, kernelVersion, state, health, capabilities, missingRequiredCapabilities, degradedCapabilities]
      properties:
        runtimeId:
          type: string
        agentId:
          type: string
        kernelVersion:
          type: string
        state:
          type: string
        health:
          type: string
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/Capability'
        missingRequiredCapabilities:
          type: array
          items:
            type: string
        degradedCapabilities:
          type: array
          items:
            type: string
    KernelEvent:
      type: object
      required: [eventId, eventType, severity, summary, sequence]
      properties:
        eventId:
          type: string
        eventType:
          type: string
        severity:
          type: string
        summary:
          type: string
        sequence:
          type: integer
          format: int32
        traceId:
          type: string
          nullable: true
    Workspace:
      type: object
      required: [workspaceId, root, branch, dirty, changedFiles]
      properties:
        workspaceId:
          type: string
        root:
          type: string
        branch:
          type: string
        dirty:
          type: boolean
        changedFiles:
          type: array
          items:
            type: string
...
    PermissionRequest:
      type: object
      required: [permissionRequestId, category, resource, sideEffectLevel, reason, status]
      properties:
        permissionRequestId:
          type: string
        category:
          type: string
        resource:
          type: string
        sideEffectLevel:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [pending, allow, deny, needs_approval]
    PatchSet:
      type: object
      required: [patchId, workspaceId, summary, operations]
      properties:
        patchId:
          type: string
        workspaceId:
          type: string
        summary:
          type: string
        operations:
          type: array
          items:
            $ref: '#/components/schemas/PatchOperation'
    PatchOperation:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [createFile, updateFile, deleteFile]
        path:
          type: string
        content:
          type: string
          nullable: true
        before:
          type: string
          nullable: true
        after:
          type: string
          nullable: true
    VerificationReport:
      type: object
      required: [reportId, verificationId, commandResults, failures]
      properties:
        reportId:
          type: string
        verificationId:
          type: string
        commandResults:
          type: array
          items:
            $ref: '#/components/schemas/CommandResult'
        failures:
          type: array
          items:
            type: string
    CommandResult:
      type: object
      required: [commandId, stdout, stderr, cancelled, timedOut]
      properties:
        commandId:
          type: string
        exitCode:
          type: integer
          format: int32
          nullable: true
        stdout:
          type: string
        stderr:
          type: string
        cancelled:
          type: boolean
        timedOut:
          type: boolean
    TerminalCommand:
      type: object
      required: [commandId, command, args, workingDirectory]
      properties:
        commandId:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        workingDirectory:
          type: string
        timeoutMs:
          type: integer
          format: int64
          nullable: true
        policyCategories:
          type: array
          items:
            type: string
    TerminalOutputChunk:
      type: object
      required: [commandId, sequence, channel, content]
      properties:
        commandId:
          type: string
        sequence:
          type: integer
          format: int64
        channel:
          type: string
          enum: [stdout, stderr, system]
        content:
          type: string
        redactionClassification:
          type: string
          nullable: true
    ReviewFinding:
      type: object
      required: [findingId, severity, filePath, message]
      properties:
        findingId:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        filePath:
          type: string
        line:
          type: integer
          format: int32
          nullable: true
        message:
          type: string
        remediation:
          type: string
          nullable: true
        missingTest:
          type: string
          nullable: true
...
    RuntimeSnapshotResponse:
      description: Runtime aggregate snapshot SdkWorkApiResponse envelope.
      allOf:
        - $ref: '#/components/schemas/SdkWorkApiResponse'
        - type: object
          required: [data]
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/SdkWorkResourceData'
                - type: object
                  properties:
                    item:
                      $ref: '#/components/schemas/RuntimeSnapshot'
...